Consolidate args and main
