+++ instant_rst/settings.py
@@ -2,16 +2,19 @@ import os
 
 ADDITIONAL_DIRS = []
 
-STATIC_DIR = os.path.join( os.path.dirname(os.path.realpath(__file__)),'templates')
-DEFAULT_FILE = os.path.join( os.path.dirname(os.path.realpath(__file__)),'templates', "index.rst")
+mod_dpath = os.path.dirname(os.path.abspath(__file__))
+
+STATIC_DIR = os.path.join(mod_dpath, 'templates')
+DEFAULT_FILE = os.path.join(mod_dpath, 'templates', "index.rst")
 HOST = "127.0.0.1"
 PORT = 5000
 URL =  "http://127.0.0.1:5000"
-BROWSER = 'firefox'
+# BROWSER = 'firefox'
+BROWSER = ''
 SECRET = 'JO34h#F*$HFHA@#&('
 
-FLASK_STATIC_FOLDER = os.path.join( os.path.dirname(os.path.realpath(__file__)),'static')
-FLASK_TEMPLATE_FOLDER = os.path.join( os.path.dirname(os.path.realpath(__file__)),'templates')
+FLASK_STATIC_FOLDER = os.path.join(mod_dpath, 'static')
+FLASK_TEMPLATE_FOLDER = os.path.join(mod_dpath, 'templates')
 
 
 _p1 = None
